Use a more general XML Formatter
