+++ lib/dradis/plugins/nexpose.rb
@@ -8,6 +8,6 @@ end
 require 'dradis/plugins/nexpose/engine'
 require 'dradis/plugins/nexpose/field_processor'
 require 'dradis/plugins/nexpose/importer'
-require 'dradis/plugins/nexpose/list_formatter'
 require 'dradis/plugins/nexpose/mapping'
 require 'dradis/plugins/nexpose/version'
+require 'dradis/plugins/nexpose/xml_formatter'
