+++ lib/nexpose/vulnerability.rb
@@ -76,14 +76,15 @@ module Nexpose
       nest = @xml.xpath("./#{method_name}").first
 
       # We need to clean up tags that have HTML content in them
+      formatter = Dradis::Plugins::Nexpose::XmlFormatter.new
       if tags_with_html_content.include?(method)
-        result = Dradis::Plugins::Nexpose::ListFormatter.new.format_list(tag)
-        result = cleanup_html(result)
+        result = formatter.format_list(tag)
+        result = formatter.cleanup_html(result)
         result = add_bc_to_ssl_cipher_list(result) if SSL_CIPHER_VULN_IDS.include?(@xml.attributes['id'].value)
         return result
       # And we need to clean up the tags with nested content in them
       elsif tags_with_nested_content.include?(method)
-        return cleanup_nested(nest)
+        return formatter.cleanup_nested(nest)
       else
         return tag
       end
@@ -106,42 +107,6 @@ module Nexpose
     def add_bc_to_ssl_cipher_list(source)
       result = source.to_s
       result.gsub!(/\n(.*?)!(.*?)/) { "\nbc. #{ $1 }!#{ $2 }\n" }
-      result
-    end
-
-    def cleanup_html(source)
-      result = source.to_s
-      result.gsub!(/<ContainerBlockElement>(.*?)<\/ContainerBlockElement>/m) { |m| "#{ $1 }" }
-      result.gsub!(/<Paragraph preformat=\"true\">(\s*)<Paragraph preformat=\"true\">(.*?)<\/Paragraph>(\s*)<\/Paragraph>/mi) do
-        text = $2
-        text[/\n/] ? "\nbc.. #{ text }\n\np. " : "@#{text}@"
-      end
-      result.gsub!(/<Paragraph preformat=\"true\">(.*?)<\/Paragraph>/mi) do
-        text = $1
-        text[/\n/] ? "\nbc.. #{ text }\n\np. " : "@#{text}@"
-      end
-      result.gsub!(/<Paragraph>(.*?)<\/Paragraph>/m) { |m| "#{ $1 }\n" }
-      result.gsub!(/<Paragraph>|<\/Paragraph>/, '')
-      result.gsub!(/          /, '')
-      result.gsub!(/   /, '')
-      result.gsub!(/\t\t/, '')
-      result.gsub!(/<URLLink(.*)LinkURL=\"(.*?)\"(.*?)>(.*?)<\/URLLink>/im) { "\"#{$4.strip}\":#{$2.strip} " }
-      result.gsub!(/<URLLink LinkTitle=\"(.*?)\"(.*?)LinkURL=\"(.*?)\"\/>/i) { "\"#{$1.strip}\":#{$3.strip} " }
-      result.gsub!(/<URLLink LinkURL=\"(.*?)\"(.*?)LinkTitle=\"(.*?)\"\/>/i) { "\"#{$3.strip}\":#{$1.strip} " }
-      result.gsub!(/&gt;/, '>')
-      result.gsub!(/&lt;/, '<')
-      result
-    end
-
-    def cleanup_nested(source)
-      result = source.to_s
-      result.gsub!(/<references>/, '')
-      result.gsub!(/<\/references>/, '')
-      result.gsub!(/<reference source=\"(.*?)\">(.*?)<\/reference>/i) { "#{$1.strip}: #{$2.strip}\n" }
-      result.gsub!(/<tags>/, '')
-      result.gsub!(/<\/tags>/, '')
-      result.gsub!(/<tag>(.*?)<\/tag>/) { "#{$1}\n" }
-      result.gsub!(/        /, '')
       result
     end
