+++ spec/xml_formatter_spec.rb
@@ -1,12 +1,12 @@
-# Run the spec by running the command: rspec spec/list_formatter_spec.rb"
+# Run the spec by running the command: rspec spec/xml_formatter_spec.rb"
 
 require 'spec_helper'
 
-describe Dradis::Plugins::Nexpose::ListFormatter do
+describe Dradis::Plugins::Nexpose::XmlFormatter do
   let(:source) { File.read(File.expand_path('../fixtures/files/lists.xml', __FILE__)) }
 
   it 'parses the <UnorderedList> and <OrderedList> elements' do
-    expect(Dradis::Plugins::Nexpose::ListFormatter.new.format_list(source)).to eq(
+    expect(Dradis::Plugins::Nexpose::XmlFormatter.new.format_list(source)).to eq(
       "* Microsoft Windows\n"\
       "## Open the Windows Control Panel.\n"\
       "## Select \"Administrative Tools\".\n"\
